feat: implement job deletion functionality and UI updates
- Added deleteJob API endpoint to backend for soft deleting transcription jobs.
- Created deleteJobResponse schema for API response.
- Implemented useDeleteJob hook in frontend to handle job deletion.
- Updated useTranscriptionManager hook to manage transcription state on deletion.
- Enhanced TranscriptionItem component with delete functionality and confirmation dialog.
- Added metrics cards to display transcription statistics.
- Created FileUploader component for uploading audio files with validation.
- Developed TranscriptionsList component to display and manage user transcriptions.
- Introduced TranscriptionStatusBadge component for displaying job statuses.
- Updated dashboard header and layout for improved user experience.
