+++ backend/app/schemas/deleteJob.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel, Field
+from typing import Optional
+
+class deleteJobResponse(BaseModel):
+    message: str
+    jobId: str
